Use Core HasVisibility in Filters HasVisibility
