+++ src/Views/Traits/Filters/HasVisibility.php
@@ -3,34 +3,13 @@
 namespace Rappasoft\LaravelLivewireTables\Views\Traits\Filters;
 
 use Rappasoft\LaravelLivewireTables\Views\{Column,Filter};
+use Rappasoft\LaravelLivewireTables\Views\Traits\Core\HasVisibility as HasCoreVisibility;
 
 trait HasVisibility
 {
-    protected bool $hiddenFromMenus = false;
-
-    protected bool $hiddenFromPills = false;
+    use HasCoreVisibility;
 
     protected bool $hiddenFromFilterCount = false;
-
-    public function isHiddenFromMenus(): bool
-    {
-        return $this->hiddenFromMenus === true;
-    }
-
-    public function isVisibleInMenus(): bool
-    {
-        return $this->hiddenFromMenus === false;
-    }
-
-    public function isHiddenFromPills(): bool
-    {
-        return $this->hiddenFromPills === true;
-    }
-
-    public function isVisibleInPills(): bool
-    {
-        return $this->hiddenFromPills === false;
-    }
 
     public function isHiddenFromFilterCount(): bool
     {
@@ -40,20 +19,6 @@ trait HasVisibility
     public function isVisibleInFilterCount(): bool
     {
         return $this->hiddenFromFilterCount === false;
-    }
-
-    public function hiddenFromMenus(): self
-    {
-        $this->hiddenFromMenus = true;
-
-        return $this;
-    }
-
-    public function hiddenFromPills(): self
-    {
-        $this->hiddenFromPills = true;
-
-        return $this;
     }
 
     public function hiddenFromFilterCount(): self
